rename rainforest.h to rfv2.h to remove confusion with v1 projects using it
